+++ src/olmo_core/data/numpy_dataset.py
@@ -37,7 +37,7 @@ from ..aliases import PathOrStr
 from ..config import Config, StrEnum
 from ..distributed.utils import barrier, get_fs_local_rank
 from ..io import _get_s3_client, get_file_size
-from .mixes import DataMixBase
+from .mixes import DataMix, DataMixBase
 from .tokenizer import TokenizerConfig
 from .utils import (
     bucket_documents,
@@ -566,6 +566,12 @@ class NumpyFSLDatasetMixture(NumpyFSLDataset):
         self._path_offset_index = path_offset_index
         self._seed = seed
 
+    @property
+    def indices_dtype(
+        self,
+    ) -> NumpyUIntTypes:
+        return np.uint32
+
     def prepare(self):
         if self.fs_local_rank == 0:
             log.info("Gathering indices...")
@@ -577,6 +583,7 @@ class NumpyFSLDatasetMixture(NumpyFSLDataset):
         sha256_hash = hashlib.sha256()
         sha256_hash.update(str(path).encode())
         sha256_hash.update(str(self._get_file_size(path)).encode())
+        sha256_hash.update(self.indices_dtype.__name__.encode())
         path_hash = sha256_hash.hexdigest()
         return (
             self.work_dir
@@ -603,18 +610,22 @@ class NumpyFSLDatasetMixture(NumpyFSLDataset):
                     max_instances = (
                         self._path_offset_index[(str(path), idx)] // self.sequence_length
                     )
-                    future = executor.submit(
-                        run_worker_func,
-                        segment_documents_into_instances,
-                        path,
-                        indices_path,
-                        max_sequence_length=self.sequence_length,
-                        eos_token_id=self.eos_token_id,
-                        dtype=self.dtype,
-                        indices_dtype=self.dtype,
-                        sample=(max_instances, self._seed),
-                    )
-                    futures.append(future)
+
+                    # Sampling from small npy files can result in 0 instance indices.
+                    # We skip processing these to avoid writing empty mmapped files.
+                    if max_instances > 0:
+                        future = executor.submit(
+                            run_worker_func,
+                            segment_documents_into_instances,
+                            path,
+                            indices_path,
+                            max_sequence_length=self.sequence_length,
+                            eos_token_id=self.eos_token_id,
+                            dtype=self.dtype,
+                            indices_dtype=self.indices_dtype,
+                            sample=(max_instances, self._seed),
+                        )
+                        futures.append(future)
 
                 concurrent.futures.wait(futures, return_when="ALL_COMPLETED")
 
@@ -625,6 +636,15 @@ class NumpyFSLDatasetMixture(NumpyFSLDataset):
                         f"Created {total_instances:,d} instances of sequence length up to "
                         f"{self.sequence_length} from '{path}'"
                     )
+
+    def _read_chunk_from_array(self, path: PathOrStr, index: int) -> torch.Tensor:
+        indices_path = self._get_indices_path(path)
+        indices = load_array_slice_into_tensor(
+            indices_path, index * 2, index * 2 + 2, self.indices_dtype
+        )
+        start_idx, end_idx = indices
+        data = load_array_slice_into_tensor(path, int(start_idx), int(end_idx), self.dtype)
+        return data
 
     def _get_file_size_and_length(
         self, path: PathOrStr, idx: int, dtype: Optional[NumpyUIntTypes] = None
@@ -1511,7 +1531,7 @@ class NumpyDatasetConfig(Config):
     """
     The paths/URLs to the numpy token ID arrays.
     """
-    mix: Optional[DataMixBase] = None
+    mix: Optional[Union[str, DataMixBase]] = None
     """
     The name of a data mix.
     """
@@ -1660,7 +1680,10 @@ class NumpyDatasetConfig(Config):
                 raise OLMoConfigurationError(
                     "Missing tokenizer identifier required to construct data mix"
                 )
-            paths, labels = self.mix.build(self.mix_base_dir, self.tokenizer.identifier)
+            mix = self.mix
+            if not isinstance(mix, DataMixBase):
+                mix = DataMix(mix)
+            paths, labels = mix.build(self.mix_base_dir, self.tokenizer.identifier)
             if metadata is None:
                 metadata = [{"label": label} for label in labels]