+++ src/olmo_core/optim/config.py
@@ -3,7 +3,19 @@ from abc import ABCMeta, abstractmethod
 from collections import OrderedDict
 from dataclasses import dataclass
 from fnmatch import fnmatch
-from typing import Any, Dict, Generic, Iterable, List, Optional, Type, TypeVar, Union
+from typing import (
+    Any,
+    Dict,
+    Generic,
+    Iterable,
+    List,
+    Optional,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+    cast,
+)
 
 import torch
 import torch.nn as nn
@@ -55,6 +67,12 @@ class OptimConfig(Config, Generic[Opt], metaclass=ABCMeta):
         You could also see unexpected behavior and very poor performance when turning this feature
         on in the middle of a run that was previously trained without compiling the optimizer
         due to the LR being restored to a float instead of a tensor.
+    """
+
+    fixed_fields: Tuple[str, ...] = ("initial_lr",)
+    """
+    These are fields that should not be overridden by the value in a checkpoint after
+    loading optimizer state.
     """
 
     @property
@@ -125,11 +143,13 @@ class OptimConfig(Config, Generic[Opt], metaclass=ABCMeta):
         kwargs = self.as_dict()
         kwargs.pop("group_overrides")
         kwargs.pop("compile")
+        kwargs.pop("fixed_fields")
 
-        optim = self.optimizer()(self.build_groups(model), **kwargs)
+        optim: torch.optim.Optimizer = self.optimizer()(self.build_groups(model), **kwargs)
 
         # Set 'lr' and 'initial_lr' in each group if needed.
-        for group in optim.param_groups:
+        fixed_fields_per_group: List[Dict[str, Any]] = [{} for _ in optim.param_groups]
+        for fixed_fields, group in zip(fixed_fields_per_group, optim.param_groups):
             lr: Optional[float] = None
             if "lr" in group:
                 lr = group["lr"]
@@ -144,8 +164,19 @@ class OptimConfig(Config, Generic[Opt], metaclass=ABCMeta):
                     group["lr"] = lr
                 group.setdefault("initial_lr", lr)
 
+            for k in self.fixed_fields:
+                if k in group:
+                    fixed_fields[k] = group[k]
+
         if self.compile:
             log.info("Compiling optimizer step...")
             optim.step = torch.compile(optim.step)
 
-        return optim
+        # Register hook to reset fixed fields after loading a checkpoint.
+        def reset_fixed_fields(opt: torch.optim.Optimizer):
+            for fixed_fields, group in zip(fixed_fields_per_group, opt.param_groups):
+                group.update(fixed_fields)
+
+        optim.register_load_state_dict_post_hook(reset_fixed_fields)
+
+        return cast(Opt, optim)